--- jdkringbuffer-core/src/main/java/org/jdkstack/jdkringbuffer/core/mpmc/version2/AbstractMpmcBlockingQueueV2.java
@@ -1,26 +1,0 @@
-package org.jdkstack.jdkringbuffer.core.mpmc.version2;
-
-import org.jdkstack.jdkringbuffer.core.AbstractLockBlockingQueueV2;
-
-/**
- * 多生产多消费MPMC阻塞队列.
- *
- * <p>线程安全处理使用CAS锁.
- *
- * @author admin
- * @param <E> e .
- */
-public abstract class AbstractMpmcBlockingQueueV2<E> extends AbstractLockBlockingQueueV2<E> {
-
-  /**
-   * This is a method description.
-   *
-   * <p>Another description after blank line.
-   *
-   * @author admin
-   * @param capacity e.
-   */
-  protected AbstractMpmcBlockingQueueV2(final int capacity) {
-    super(capacity);
-  }
-}
