+++ jdkringbuffer-core/src/main/java/org/jdkstack/jdkringbuffer/core/mpmc/version2/MpmcBlockingQueueV2.java
@@ -1,16 +1,17 @@
 package org.jdkstack.jdkringbuffer.core.mpmc.version2;
 
+import org.jdkstack.jdkringbuffer.core.AbstractLockBlockingQueueV2;
 import org.jdkstack.jdkringbuffer.core.Constants;
 
 /**
- * This is a class description.
+ * 多生产多消费MPMC阻塞队列.
  *
- * <p>Another description after blank line.
+ * <p>线程安全处理使用CAS锁.
  *
  * @author admin
- * @param <E> e.
+ * @param <E> e .
  */
-public class MpmcBlockingQueueV2<E> extends AbstractMpmcBlockingQueueV2<E> {
+public class MpmcBlockingQueueV2<E> extends AbstractLockBlockingQueueV2<E> {
 
   /**
    * This is a method description.
