+++ jdkringbuffer-core/src/main/java/org/jdkstack/jdkringbuffer/core/mpsc/version1/MpscBlockingQueueV1.java
@@ -1,5 +1,6 @@
 package org.jdkstack.jdkringbuffer.core.mpsc.version1;
 
+import org.jdkstack.jdkringbuffer.core.AbstractLockBlockingQueueV1;
 import org.jdkstack.jdkringbuffer.core.Constants;
 
 /**
@@ -10,7 +11,7 @@ import org.jdkstack.jdkringbuffer.core.Constants;
  * @author admin
  * @param <E> e.
  */
-public class MpscBlockingQueueV1<E> extends AbstractMpscBlockingQueueV1<E> {
+public class MpscBlockingQueueV1<E> extends AbstractLockBlockingQueueV1<E> {
 
   /**
    * This is a method description.
@@ -33,5 +34,31 @@ public class MpscBlockingQueueV1<E> extends AbstractMpscBlockingQueueV1<E> {
    */
   public MpscBlockingQueueV1(final int capacity) {
     super(capacity);
+  }
+
+  /**
+   * This is a method description.
+   *
+   * <p>Another description after blank line.
+   *
+   * @author admin
+   * @return E e.
+   */
+  @Override
+  public E poll() {
+    final int headSeq = this.head.get();
+    E e = null;
+    // 检查队列中是否有元素.   检查是否被其他线程修改,如果返回true,则没有修改,可以更新值.
+    if (0 > headSeq || this.tail.get() > headSeq) {
+      // poll计数.
+      this.head.getAndIncrement();
+      // 获取环形数组索引位置.
+      final int bufferIndex = headSeq & index;
+      // 根据索引位置获取元素.
+      e = this.ringBuffer[bufferIndex];
+      // 将索引位置设置为空.
+      this.ringBuffer[bufferIndex] = null;
+    }
+    return e;
   }
 }
