+++ jdkringbuffer-core/src/main/java/org/jdkstack/jdkringbuffer/core/mpsc/version2/MpscBlockingQueueV2.java
@@ -1,6 +1,8 @@
 package org.jdkstack.jdkringbuffer.core.mpsc.version2;
 
+import org.jdkstack.jdkringbuffer.core.AbstractLockBlockingQueueV2;
 import org.jdkstack.jdkringbuffer.core.Constants;
+import org.jdkstack.jdkringbuffer.core.Entry;
 
 /**
  * This is a class description.
@@ -10,7 +12,7 @@ import org.jdkstack.jdkringbuffer.core.Constants;
  * @author admin
  * @param <E> e.
  */
-public class MpscBlockingQueueV2<E> extends AbstractMpscBlockingQueueV2<E> {
+public class MpscBlockingQueueV2<E> extends AbstractLockBlockingQueueV2<E> {
 
   /**
    * This is a method description.
@@ -33,5 +35,29 @@ public class MpscBlockingQueueV2<E> extends AbstractMpscBlockingQueueV2<E> {
    */
   public MpscBlockingQueueV2(final int capacity) {
     super(capacity);
+  }
+
+  /**
+   * This is a method description.
+   *
+   * <p>Another description after blank line.
+   *
+   * @author admin
+   * @return boolean e.
+   */
+  @Override
+  public E poll() {
+    final int headSeq = this.head.get();
+    final Entry<E> cell = buffer[headSeq & index];
+    final int seq = cell.getSeq();
+    final int dif = seq - (headSeq + 1);
+    E e = null;
+    if (dif == 0) {
+      e = cell.getEntry();
+      cell.setEntry(null);
+      cell.setSeq(headSeq + index + 1);
+      this.head.getAndIncrement();
+    }
+    return e;
   }
 }
